requirements.txt
antlr4-python3-runtime==4.9.3
colorama==0.4.4
